Updated database schema
Ergast updated one of their relations to record the time and date of qualifying and sprint events in addition to the race itself. This branch adds an update to the schema for that change. Note that any new data cannot be added to the old database because of this. I am planning on making a script that automatically processes the CSV files. The CSVs that were manually processed and in the old schema will be in the "OLD ergast-database-csv" folder if needed.