New database schema, connects to user database automatically instead of manually creating temp database, loading csv using pandas (#16)
* Connect to user database to create ergast database

Allows for the script to work on its own so a manual/temporary postgress database isn't necessary.
Connects to the user database created by default to create the database instead.

* Updated database schema

Ergast updated one of their relations to record the time and date of qualifying and sprint events in addition to the race itself. This branch adds an update to the schema for that change. Note that any new data cannot be added to the old database because of this. I am planning on making a script that automatically processes the CSV files. The CSVs that were manually processed and in the old schema will be in the "OLD ergast-database-csv" folder if needed.